backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.0

# Data validation
pydantic==2.7.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9    # For Postgres (optional; not used in SQLite)
python-dotenv==1.0.1      # For .env config

# HTTP client
httpx==0.27.2

# Matching (TF-IDF, cosine similarity)
scikit-learn==1.5.0

# Optional utilities
numpy==1.26.4              # scikit-learn dependency (sometimes needs pinning)

# Optional ML/DB (add when needed)
# pandas==2.2.3
# alembic==1.14.0
